refactored models and search class into individual submodules
